HlsDataSource change constructor type to size_t
android 32 bit builds can't pass a uint64_t to a size_t.

Bug: 1266991
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4742907
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1181215}

+++ media/filters/hls_data_source_provider.h
@@ -42,7 +42,7 @@ class MEDIA_EXPORT HlsDataSource {
   using ReadStatus = TypedStatus<ReadStatusTraits>;
   using ReadCb = base::OnceCallback<void(ReadStatus::Or<size_t>)>;
 
-  explicit HlsDataSource(absl::optional<uint64_t> size) : size_(size) {}
+  explicit HlsDataSource(absl::optional<size_t> size) : size_(size) {}
   virtual ~HlsDataSource();
 
   // Issues a read to the underlying data source, writing the results to
